--- src/context/mod.rs
@@ -1,9 +1,0 @@
-#![allow(missing_docs)]
-
-mod endpoint;
-mod task;
-mod responder;
-
-pub use self::endpoint::EndpointContext;
-pub use self::task::TaskContext;
-pub use self::responder::ResponderContext;
