+++ src/response/mod.rs
@@ -1,5 +1,6 @@
 #![allow(missing_docs)]
 
+mod context;
 mod responder;
 mod response;
 
@@ -7,7 +8,6 @@ pub use hyper::header;
 pub use hyper::mime;
 pub use hyper::StatusCode;
 
+pub use self::context::ResponderContext;
 pub use self::responder::{IntoResponder, Responder};
 pub use self::response::{Response, ResponseBuilder};
-
-pub use context::ResponderContext;
